requirements.txt
pandas==1.4.4
pdfminer.six==20221105
pdfplumber==0.9.0
PyMuPDF==1.22.3
python_docx==0.8.11
PyPDF2==3.0.1